[Hotfix] Reap shared memory on exit (#7)
* config

* configure shmem by arg

* configure thread pool with arg

* ;

* ;

* reap shmem

* misc

* perm error

* .

+++ examples/asplos25/run.sh
@@ -53,8 +53,15 @@ DATA_ARGS="
     --split 949,50,1
 "
 
-if [ -f "${FUSED_KERNEL_LOCK}" ]; then
+if [ -n "$FUSED_KERNEL_LOCK" ] && [ -f "${FUSED_KERNEL_LOCK}" ]; then
     rm ${FUSED_KERNEL_LOCK}
+fi
+
+if [ -n "${SPIRAL_SHMEM_NAME}" ] && [ -e "/dev/shm${SPIRAL_SHMEM_NAME}" ]; then
+    if [ ! -r "/dev/shm${SPIRAL_SHMEM_NAME}" ] || [ ! -w "/dev/shm${SPIRAL_SHMEM_NAME}" ]; then
+        echo "Permission error: /dev/shm${SPIRAL_SHMEM_NAME} exists already and is not readable/writable"
+        exit 1
+    fi
 fi
 
 EXEC_CMD="python ${MEGATRON_PATH}/pretrain_gpt.py ${EXTRA_ARGS} ${DISTRIBUTED_ARGS} ${GPT_ARGS} ${DATA_ARGS}"
